Statistics: traverse tests refactoring
