+++ ydb/core/statistics/aggregator/ut/ut_analyze_columnshard.cpp
@@ -9,44 +9,13 @@
 namespace NKikimr {
 namespace NStat {
 
-struct TTableInfo {
-    std::vector<ui64> ShardIds;
-    ui64 SaTabletId; 
-    TPathId DomainKey;
-    TPathId PathId;
-};
 
-std::vector<TTableInfo> CreateDatabaseTables(TTestEnv& env, ui8 tableCount, ui8 shardCount) {
-    auto init = [&] () {
-        CreateDatabase(env, "Database");
-        for (ui8 tableId = 1; tableId <= tableCount; tableId++) {
-            CreateColumnStoreTable(env, "Database", Sprintf("Table%u", tableId), shardCount);
-        }
-    };
-    std::thread initThread(init);
-
-    auto& runtime = *env.GetServer().GetRuntime();
-    auto sender = runtime.AllocateEdgeActor();
-
-    runtime.SimulateSleep(TDuration::Seconds(10));
-    initThread.join();
-
-    std::vector<TTableInfo> ret;
-    for (ui8 tableId = 1; tableId <= tableCount; tableId++) {
-        TTableInfo tableInfo;
-        const TString path = Sprintf("/Root/Database/Table%u", tableId);
-        tableInfo.ShardIds = GetColumnTableShards(runtime, sender, path);
-        tableInfo.PathId = ResolvePathId(runtime, path, &tableInfo.DomainKey, &tableInfo.SaTabletId);
-        ret.emplace_back(tableInfo);
-    }
-    return ret;
-}    
 
 Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeOneColumnTable) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
 
         AnalyzeTable(runtime, tableInfo.ShardIds[0], tableInfo.PathId);
 
@@ -56,7 +25,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeAnalyzeOneColumnTableSpecificColumns) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
 
         Analyze(runtime, tableInfo.SaTabletId, {{tableInfo.PathId, {1, 2}}});
     }
@@ -64,7 +33,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeTwoColumnTables) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfos = CreateDatabaseTables(env, 2, 1);
+        auto tableInfos = CreateDatabaseColumnTables(env, 2, 1);
 
         Analyze(runtime, tableInfos[0].SaTabletId, {tableInfos[0].PathId, tableInfos[1].PathId});
     }
@@ -78,7 +47,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
             ev.Reset();
         });
 
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
 
         const TString operationId = "operationId";
 
@@ -100,8 +69,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeSameOperationId) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
         const TString operationId = "operationId";
 
@@ -122,8 +90,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaBeforeAnalyzeTableResponse) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
 
         bool eventSeen = false;
@@ -146,8 +113,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaBeforeResolve) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
 
         int observerCount = 0;
@@ -170,8 +136,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaBeforeReqDistribution) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
 
         bool eventSeen = false;
@@ -194,8 +159,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaBeforeAggregate) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
 
         bool eventSeen = false;
@@ -218,8 +182,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaBeforeSave) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 1)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 1)[0];
         auto sender = runtime.AllocateEdgeActor();
 
         bool eventSeen = false;
@@ -243,8 +206,7 @@ Y_UNIT_TEST_SUITE(AnalyzeColumnshard) {
     Y_UNIT_TEST(AnalyzeRebootSaInAggregate) {
         TTestEnv env(1, 1);
         auto& runtime = *env.GetServer().GetRuntime();
-        auto tableInfo = CreateDatabaseTables(env, 1, 10)[0];
-
+        auto tableInfo = CreateDatabaseColumnTables(env, 1, 10)[0];
         auto sender = runtime.AllocateEdgeActor();
         
         int observerCount = 0;
